Refactor Dashboard Controls and WidgetServiceHelper (#1222)
