+++ tools/Dashboard/DevHome.Dashboard/Controls/WidgetBoard.cs
@@ -7,7 +7,7 @@ using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Windows.Foundation;
 
-namespace DevHome.Dashboard.Views;
+namespace DevHome.Dashboard.Controls;
 public sealed class WidgetBoard : Panel
 {
     private double _columnWidth;
