+++ tools/Dashboard/DevHome.Dashboard/Controls/WidgetControl.xaml
@@ -2,7 +2,7 @@
 <!-- Licensed under the MIT License. -->
 
 <UserControl
-    x:Class="DevHome.Dashboard.Views.WidgetControl"
+    x:Class="DevHome.Dashboard.Controls.WidgetControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:views="using:DevHome.Dashboard.Views"
