+++ tools/Dashboard/DevHome.Dashboard/Views/DashboardView.xaml
@@ -11,6 +11,7 @@
     xmlns:commonviews="using:DevHome.Common.Views"
     xmlns:vm="using:DevHome.Dashboard.ViewModels"
     xmlns:views="using:DevHome.Dashboard.Views"
+    xmlns:controls="using:DevHome.Dashboard.Controls"
     xmlns:behaviors="using:DevHome.Common.Behaviors"
     behaviors:NavigationViewHeaderBehavior.HeaderMode="Never"
     mc:Ignorable="d">
@@ -63,12 +64,12 @@
                           SelectionMode="None">
                     <GridView.ItemTemplate>
                         <DataTemplate x:DataType="vm:WidgetViewModel">
-                            <views:WidgetControl WidgetSource="{x:Bind}" />
+                            <controls:WidgetControl WidgetSource="{x:Bind}" />
                         </DataTemplate>
                     </GridView.ItemTemplate>
                     <GridView.ItemsPanel>
                         <ItemsPanelTemplate>
-                            <views:WidgetBoard />
+                            <controls:WidgetBoard />
                         </ItemsPanelTemplate>
                     </GridView.ItemsPanel>
                     <GridView.ItemContainerStyle>
@@ -79,13 +80,14 @@
                     </GridView.ItemContainerStyle>
                 </GridView>
 
-              <!-- Loading Progress -->
-              <ProgressRing x:Name="LoadingWidgetsProgressRing"
-                            Margin="0,150,0,0"
-                            IsActive="True"
-                            HorizontalAlignment="Center"/>
+                <!-- Widget grid loading progress -->
+                <ProgressRing x:Name="LoadingWidgetsProgressRing"
+                              Margin="0,150,0,0"
+                              IsActive="True"
+                              HorizontalAlignment="Center"/>
 
-              <StackPanel x:Name="NoWidgetsStackPanel" HorizontalAlignment="Center" Margin="0,150,0,0" Visibility="Collapsed">
+                <!-- Widget grid empty message -->
+                <StackPanel x:Name="NoWidgetsStackPanel" HorizontalAlignment="Center" Margin="0,150,0,0" Visibility="Collapsed">
                     <TextBlock x:Uid="NoWidgetsAdded" HorizontalAlignment="Center" />
                     <HyperlinkButton x:Name="AddFirstWidgetLink" x:Uid="AddFirstWidgetLink" Click="AddWidget_Click" HorizontalAlignment="Center" />
                 </StackPanel>
